Move getImage from balance api endpoint to coinmarketcap lib
